requirements.txt
Flask==2.3.3
python-dotenv==1.1.1
pydantic
pypdf2==3.0.1
langchain==0.3.25
langchain-google-genai==2.0.1
chargebee
gunicorn